Undefined values.
(define ...) and (set! ...) have undefined return values.  If a
top-level expression evaluates to an undefined value nothing is printed.
If an "if" condition or an argument to "cons" is undefined there's an
error.  Ideally there'd be an error any time an expression evaluates to
an undefined value in a context where the value is not thrown away.

The reason this is implemented with a placeholder value is to allow
stuff like this:

  ; This prints nothing even though "if" normally returns a value:
  (if foo
    (define bar 'bar)
    (define baz 'baz))